normal int
Source: https://github.com/msupply-foundation/open-msupply/commit/81ff8436ec4281537aaa16b4773bb18d127da6de

+++ server/graphql/general/src/queries/sync_status.rs
@@ -13,7 +13,7 @@ use crate::sync_api_error::SyncErrorNode;
 
 pub struct SyncStatusNode {
     started: NaiveDateTime,
-    duration_in_seconds: i64,
+    duration_in_seconds: i32,
     finished: Option<NaiveDateTime>,
 }
 
@@ -23,7 +23,7 @@ impl SyncStatusNode {
         DateTime::<Utc>::from_naive_utc_and_offset(self.started, Utc)
     }
 
-    async fn duration_in_seconds(&self) -> i64 {
+    async fn duration_in_seconds(&self) -> i32 {
         self.duration_in_seconds
     }
